Develop

+++ CLImageEditor/ImageTools/CLEffectTool/CLEffect/CLPixellateEffect.m
@@ -80,8 +80,8 @@
     CFDataRef m_DataRef = CGDataProviderCopyData(CGImageGetDataProvider(inImage));
     UInt8 * m_PixelBuf = (UInt8 *) CFDataGetBytePtr(m_DataRef);
     
-    int width = CGImageGetWidth(inImage);
-    int height = CGImageGetHeight(inImage);
+    int width  = (int)CGImageGetWidth(inImage);
+    int height = (int)CGImageGetHeight(inImage);
     
     BOOL breakOut = NO;
     for (int x = 0;breakOut==NO && x < width; ++x) {
